pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.6",
    "mcp[cli]>=1.20.0",
    "motor>=3.7.1",
    "pymongo>=4.16.0",
    "requests>=2.32.5",
]
